refactor: put the fetch function into a gallery manager
put the fetch function into a gallery manager

MODULE-1-PART-3-gallery
