+++ gallery/gallery.manager.js
@@ -1,12 +1,13 @@
 import { wrapNumbersInHtml, wrapUrlsInHtml } from "../utilities/htmlWrapping/htmlWrapping.js";
-import { fetchImages } from "../utilities/dataSubmitting/fetch.js";
 import { UrlManipulationService } from "../services/url-manipulation.service.js";
+import { BASE_URL } from "../data/constants.js";
 
 export class GalleryManager {
 
     constructor() {
         this.service = new UrlManipulationService();
     }
+
     renderPagesList(totalNumberOfPages) {
         const pages = document.getElementById('pages');
         pages.innerHTML = wrapNumbersInHtml(totalNumberOfPages);
@@ -19,9 +20,19 @@ export class GalleryManager {
 
     async getImages() {
         const pageNumber = this.service.getPageNumberFromUrl();
-        const response = await fetchImages(pageNumber);
+        const response = await this.fetchImages(pageNumber);
         return response;
     }
 
+    async fetchImages(pageNumber) {
+        const accessToken = localStorage.getItem('token');
+        const url = `${BASE_URL}/gallery?page=${pageNumber}`;
 
+        return await fetch(url, {
+            method: 'GET',
+            headers: {
+                Authorization: accessToken
+            }
+        });
+    }
 }
